Cleanup and Planning for the journey
Developing a bootloader for x86-64 system is not at all an easy task. The code at this point of stable, but nothing boots as far as I know. As this project is also my major project => I need to have some presentable results, therefore, the development of the operating system will be done for the architectures in the given order : 16-bit/real, 32-bit/protected, 64-bit/long, ARM, RISC-V. Five separate folders. That's the plan.