getchar implemented and it works!!
will work in conjugation with 'shelling' prgrm to provide a prompt to the user. Note that for some keys like PG DN, PG UP, and ENTER, only the scan codes are being returned because the way the functionality of these keys is not only related to modifying text on a line, but also to move b/w pages and all. Maybe I will end up with a 'noodled' code, but it will be mine and it will be something. Will optimize it later 'Premature optimization is the root of all evil.' - D. Knuth