Time, Date, Lexer WORKING!!
The error which was preventing the code from successfully was quite stupid. I never expected that the operating system will grow more than a kilobyte in size, so during booting, I was reading only 2 sectors = 1 KB. However, as I added more functionality, the size grew (currently 1.55 KB), and as only 1 KB of kernel code was being loaded, errors happened when a label in size of kernel - 1 KB part of the kernel was accessed. Now I am loading 16 sectors = 8 KB.