New nomenclature, strcpy and disk_com
1. New nomenclature for command processing files
There are two types of files: one which contain subroutines to deal with
devices, and second which deal process the commands using core
subroutines. First class of subroutines will be clubbed into a file
having the same name as the device/process it deals with. Second class
of subroutines will be put into files having the same name as the "core"
file but suffixed with "_com". Both of these will be put in a folder
having the same name as the device/process they deal with.
Example: Core subroutines dealing with Advanced Power Management (APM)
are put in apm.asm, and subroutines processing "power" command are put
in apm_com.asm file. Both of these files are put in a folder named apm.

2. strcpy has been implemented. Lexer code has been generalized to deal
   with a variety of separators and has been moved to include/string/.

3. disk status, read and write commands have been implemented. They are
   found to be working correctly. :-)