'run ssss:oooo' command is working.
In disk2.bin, a file asm.bin has been included. It contains binary code
to put a red color 'X' on the top left of the screen (0xb8000). Using
disk read, the first sector of disk2 was read and loaded at a memory
location. Then using call far we jump to that location, run the code and
return back. After returning, we reset the value of segment registers
CS, DS, and ES in case they have been altered.